refactor #81: Todo interface 디렉터리 분리
