+++ client/src/interfaces/index.ts
@@ -1,7 +1,3 @@
-export interface QueryData {
-  data: Todo[];
-}
-
 export interface Todo {
   title: string;
   userId: string;
