+++ client/src/mutations/useCreateTodoMutation/api.ts
@@ -1,7 +1,7 @@
 import { api } from '@api/index';
-import { RESOURCE } from '@queries/useTodos/api';
+import { RESOURCE } from '@queries/useTodosQuery/api';
 
-import type { Todo } from '@queries/useTodos/types';
+import type { Todo } from '@interfaces/index';
 import type { TodoFormInput } from './types';
 
 export const createTodo = async (data: TodoFormInput) =>
