+++ client/src/queries/index.ts
@@ -1,2 +1,2 @@
 export { default as useUser } from './useUserQuery';
-export { default as useTodos } from './useTodos';
+export { default as useTodos } from './useTodosQuery';
