+++ client/src/queries/useTodosQuery/api.ts
@@ -1,10 +1,10 @@
 import { api } from '@api/index';
-import type { QueryData } from './types';
+import type { Todo } from '@interfaces/index';
 
 export const RESOURCE = '/todos';
 
 export const getTodos = async (userId?: string) =>
-  api.get<QueryData>(`${RESOURCE}`, { params: { userId } });
+  api.get<Todo[]>(`${RESOURCE}`, { params: { userId } });
 
 export const deleteTodo = async (id: string) =>
   api.delete<null>(`${RESOURCE}/${id}`);
