+++ client/tsconfig.paths.json
@@ -6,6 +6,7 @@
       "@assets/*": ["./assets/*"],
       "@components/*": ["./components/*"],
       "@hooks/*": ["./hooks/*"],
+      "@interfaces/*": ["./interfaces/*"],
       "@mutations/*": ["./mutations/*"],
       "@pages/*": ["./pages/*"],
       "@queries/*": ["./queries/*"],
